requirements.txt
requests
gmssl
pycryptodome
lxml
regex
pytesseract
pillow